+++ spec/factory.rb
@@ -139,7 +139,7 @@ module AIXM
           name: 'VOR NAVAID',
           xy: AIXM.xy(lat: %q(47°51'33"N), long: %q(007°33'36"E)),
           z: AIXM.z(500, :qnh),
-          type: :VOR,
+          type: :vor,
           f: AIXM.f(111, :mhz),
           north: :geographic
         ).tap do |vor|