+++ spec/lib/aixm/feature/navigational_aid/designated_point_spec.rb
@@ -8,20 +8,20 @@ describe AIXM::Feature::NavigationalAid::DesignatedPoint do
 
     describe :kind do
       it "must return class or type" do
-        subject.kind.must_equal :ICAO
+        subject.kind.must_equal "DesignatedPoint:ICAO"
       end
     end
 
     describe :to_digest do
       it "must return digest of payload" do
-        subject.to_digest.must_equal 582056647
+        subject.to_digest.must_equal 755263208
       end
     end
 
     describe :to_xml do
       it "must build correct XML of VOR with OFM extension" do
         subject.to_xml(:ofm).must_equal <<~END
-          <!-- Navigational aid: [ICAO] DESIGNATED POINT NAVAID -->
+          <!-- Navigational aid: [DesignatedPoint:ICAO] DESIGNATED POINT NAVAID -->
           <Dpn>
             <DpnUid newEntity="true">
               <codeId>DPN</codeId>